pyproject.toml
[tool.poetry]
name = "fastapi-sample"
version = "0.1.0"
description = ""
authors = ["lyingtakemura <106205052+lyingtakemura@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.2"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.29"
pytest = "^8.1.1"
httpx = "^0.27.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
